apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.9
pdfplumber==0.11.4
pypdf==4.3.1
python-docx==1.1.2
